refactor(Utils): move barcode stuff to a dedicated file
