+++ src/views/PriceValidationAssistant.vue
@@ -49,6 +49,7 @@ import { defineAsyncComponent } from 'vue'
 import constants from '../constants'
 import api from '../services/api'
 import { useAppStore } from '../store'
+import barcode_utils from '../utils/barcode.js'
 import utils from '../utils.js'
 
 export default {
@@ -172,7 +173,7 @@ export default {
             if (data.items[i]['predictions'].length > 0) {
               const priceTagPrediction = data.items[i]['predictions'][0]
               const label = priceTagPrediction['data']
-              const barcodeString = label.barcode ? utils.cleanBarcode(label.barcode.toString()) : ''
+              const barcodeString = label.barcode ? barcode_utils.cleanBarcode(label.barcode.toString()) : ''
 
               // The first schema was not versioned, so if the field is missing,
               // we assume it's schema version 1.0
